place the runs in their own filesystem
This places all runs in their own filesystem so that later we can simply
destroy the FS. This saves me from having to go through a list of all
filesystems to determine if they are the runs, the corpus, or the logs.
It will also make it easier to see space accounting.