Rename the sendfile
Use a more generic name for the sendfile to be received during tests.

During my initial testing I made 3 sendfiles of varying size, and would
move a symbolic link called "100GB.zfs" between them. This was pretty
dumb. I get the size of the sendfile later, and we are even considering
that a user may not completely ingest the file to get the average
speed.