Adds linux_get_current_txg
This is a really fast way to get the current TXG on linux. I do not
think the other implementations have something like this, so the method
call lets you know its linux specific.

I could not use the generic zdb call for the txg tracker because it
took longer than a second to produce a result.